refactor: move promptComposeField to executor_messages_prompt.go
